Using scaffolded EF classes

+++ Ianf.Fittrack.Workouts/Domain/Weight.cs
@@ -7,16 +7,16 @@ namespace Ianf.Fittrack.Workouts.Domain
 {
     public struct Weight : IEquatable<Weight>
     {
-        public double Value { get; }
+        public decimal Value { get; }
 
-        private Weight(double weight) => Value = weight;
+        private Weight(decimal weight) => Value = weight;
 
-        public static Option<Weight> CreateWeight(double weight) =>
+        public static Option<Weight> CreateWeight(decimal weight) =>
             IsValid(weight)
                 ? Some(new Weight(weight))
                 : None;
 
-        private static bool IsValid(double weight) => weight > 0 && (weight % 0.5) == 0;
+        private static bool IsValid(decimal weight) => weight > 0 && (weight % 0.5M) == 0;
 
         public bool Equals(Weight other) => Value == other.Value;
     }
